Memory saving: restrict some types to uint8 (e.g. ChannelType, TreeType, PredMode, ...). Note: ChromaFormat has been left apart since it would cause a crash in simdInterpolateLuma10Bit2P16 (to be investigated).

+++ source/Lib/CommonLib/IntraPrediction.cpp
@@ -263,7 +263,7 @@ void IntraPrediction::predIntraAng( const ComponentID compId, PelBuf &piPred, co
   }
 }
 
-void IntraPrediction::predIntraChromaLM(const ComponentID compID, PelBuf &piPred, const PredictionUnit &pu, const CompArea& chromaArea, int intraDir)
+void IntraPrediction::predIntraChromaLM(const ComponentID compID, PelBuf &piPred, const PredictionUnit &pu, const CompArea &chromaArea, uint8_t intraDir)
 {
   int  iLumaStride = 0;
   PelBuf Temp;
