+++ source/Lib/CommonLib/UnitTools.h
@@ -130,11 +130,11 @@ namespace PU
   int  getIntraMPMs(const PredictionUnit &pu, unsigned *mpm, const ChannelType &channelType = CHANNEL_TYPE_LUMA);
   bool          isMIP                 (const PredictionUnit &pu, const ChannelType &chType = CHANNEL_TYPE_LUMA);
   bool          isDMChromaMIP         (const PredictionUnit &pu);
-  uint32_t      getIntraDirLuma       (const PredictionUnit &pu);
+  uint8_t       getIntraDirLuma       (const PredictionUnit &pu);
   void getIntraChromaCandModes        (const PredictionUnit &pu, unsigned modeList[NUM_CHROMA_MODE]);
   const PredictionUnit &getCoLocatedLumaPU(const PredictionUnit &pu);
-  uint32_t getFinalIntraMode              (const PredictionUnit &pu, const ChannelType &chType);
-  uint32_t getCoLocatedIntraLumaMode      (const PredictionUnit &pu);
+  uint8_t getFinalIntraMode               (const PredictionUnit &pu, const ChannelType &chType);
+  uint8_t getCoLocatedIntraLumaMode       (const PredictionUnit &pu);
   int  getWideAngle                   ( const TransformUnit &tu, const uint32_t dirMode, const ComponentID compID );
   void getInterMergeCandidates        (const PredictionUnit &pu, MergeCtx& mrgCtx,
     int mmvdList,
